DSND-541: Generate and set etag on incoming charges externaldata
