+++ pom.xml
@@ -28,6 +28,7 @@
 		<docker-maven-plugin.version>1.0.0</docker-maven-plugin.version>
 		<structured-logging.version>1.9.12</structured-logging.version>
 		<api-sdk-manager-java-library.version>1.0.4</api-sdk-manager-java-library.version>
+		<api-helper-java-library.version>1.4.5</api-helper-java-library.version>
 		<cucumber-java.version>7.2.3</cucumber-java.version>
 		<jib-maven-plugin.version>3.1.1</jib-maven-plugin.version>
 		<skip.integration.tests>false</skip.integration.tests>
@@ -153,6 +154,11 @@
 			<groupId>uk.gov.companieshouse</groupId>
 			<artifactId>api-sdk-manager-java-library</artifactId>
 			<version>${api-sdk-manager-java-library.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>uk.gov.companieshouse</groupId>
+			<artifactId>api-helper-java</artifactId>
+			<version>${api-helper-java-library.version}</version>
 		</dependency>
 	</dependencies>
