+++ src/main/java/uk/gov/companieshouse/charges/data/config/ApplicationConfig.java
@@ -1,11 +1,22 @@
 package uk.gov.companieshouse.charges.data.config;
 
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.databind.module.SimpleModule;
+import java.time.LocalDate;
+import java.util.List;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Scope;
+import org.springframework.data.mongodb.core.convert.MongoCustomConversions;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import uk.gov.companieshouse.api.InternalApiClient;
+import uk.gov.companieshouse.charges.data.converter.ChargeApiReadConverter;
+import uk.gov.companieshouse.charges.data.converter.ChargeApiWriteConverter;
+import uk.gov.companieshouse.charges.data.serialization.LocalDateDeSerializer;
+import uk.gov.companieshouse.charges.data.serialization.LocalDateSerializer;
 import uk.gov.companieshouse.environment.EnvironmentReader;
 import uk.gov.companieshouse.environment.impl.EnvironmentReaderImpl;
 import uk.gov.companieshouse.sdk.manager.ApiSdkManager;
@@ -24,5 +35,31 @@ public class ApplicationConfig implements WebMvcConfigurer {
         return ApiSdkManager.getPrivateSDK();
     }
 
+    /**
+     * mongoCustomConversions.
+     *
+     * @return MongoCustomConversions.
+     */
+    @Bean
+    public MongoCustomConversions mongoCustomConversions() {
+        ObjectMapper objectMapper = mongoDbObjectMapper();
+        return new MongoCustomConversions(List.of(new ChargeApiWriteConverter(objectMapper),
+                new ChargeApiReadConverter(objectMapper)));
+    }
 
+    /**
+     * Mongo DB Object Mapper.
+     *
+     * @return ObjectMapper.
+     */
+    private ObjectMapper mongoDbObjectMapper() {
+        ObjectMapper objectMapper = new ObjectMapper();
+        objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+        objectMapper.configure(SerializationFeature.FAIL_ON_EMPTY_BEANS, false);
+        SimpleModule module = new SimpleModule();
+        module.addSerializer(LocalDate.class, new LocalDateSerializer());
+        module.addDeserializer(LocalDate.class, new LocalDateDeSerializer());
+        objectMapper.registerModule(module);
+        return objectMapper;
+    }
 }
